backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Word document translation via OpenAI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "pydantic-settings>=2.13.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.0",
]
